Refactor layout structure; introduce Ihateserverside component for sidebar management and audio player integration; remove unused metadata file and clean up AudioPlayerContext
