+++ app/components/AudioPlayerContext.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { createContext, useContext, useState } from 'react';
 
 interface Track {
